+++ src/AddressRegistry.Api.Legacy/CrabHouseNumber/CrabHouseNumberQuery.cs
@@ -17,10 +17,13 @@ namespace AddressRegistry.Api.Legacy.CrabHouseNumber
         protected override IQueryable<CrabIdToPersistentLocalIdItem> Filter(FilteringHeader<CrabHouseNumberAddressFilter> filtering)
         {
             var query = _context.CrabIdToPersistentLocalIds
-                .Where(x => !x.IsRemoved && x.HouseNumberId.HasValue && x.PersistentLocalId.HasValue);
+                .Where(x => x.HouseNumberId.HasValue && x.PersistentLocalId.HasValue);
 
-            if (filtering.ShouldFilter && filtering.Filter.CrabHouseNumberId.HasValue)
-                query = query.Where(x => x.HouseNumberId == filtering.Filter.CrabHouseNumberId);
+            var parsed = int.TryParse(filtering.Filter?.CrabHouseNumberId, out var objectId);
+            if (filtering.ShouldFilter && parsed)
+                query = query.Where(x => x.HouseNumberId == objectId);
+            else if (!parsed && !string.IsNullOrEmpty(filtering.Filter?.CrabHouseNumberId))
+                return new List<CrabIdToPersistentLocalIdItem>().AsQueryable();
 
             return query;
         }
@@ -40,6 +43,6 @@ namespace AddressRegistry.Api.Legacy.CrabHouseNumber
 
     public class CrabHouseNumberAddressFilter
     {
-        public int? CrabHouseNumberId { get; set; }
+        public string CrabHouseNumberId { get; set; }
     }
 }