+++ src/AddressRegistry.Api.Legacy/CrabSubaddress/CrabSubaddressQuery.cs
@@ -17,10 +17,13 @@ namespace AddressRegistry.Api.Legacy.CrabSubaddress
         protected override IQueryable<CrabIdToPersistentLocalIdItem> Filter(FilteringHeader<CrabSubaddressAddressFilter> filtering)
         {
             var query = _context.CrabIdToPersistentLocalIds
-                .Where(x => !x.IsRemoved && x.SubaddressId.HasValue && x.PersistentLocalId.HasValue);
+                .Where(x => x.SubaddressId.HasValue && x.PersistentLocalId.HasValue);
 
-            if (filtering.ShouldFilter && filtering.Filter.CrabSubaddressId.HasValue)
-                query = query.Where(x => x.SubaddressId == filtering.Filter.CrabSubaddressId);
+            var parsed = int.TryParse(filtering.Filter?.CrabSubaddressId, out var objectId);
+            if (filtering.ShouldFilter && parsed)
+                query = query.Where(x => x.SubaddressId == objectId);
+            else if (!parsed && !string.IsNullOrEmpty(filtering.Filter?.CrabSubaddressId))
+                return new List<CrabIdToPersistentLocalIdItem>().AsQueryable();
 
             return query;
         }
@@ -40,6 +43,6 @@ namespace AddressRegistry.Api.Legacy.CrabSubaddress
 
     public class CrabSubaddressAddressFilter
     {
-        public int? CrabSubaddressId { get; set; }
+        public string CrabSubaddressId { get; set; }
     }
 }