+++ src/AddressRegistry.Projections.Legacy/LegacyContext.cs
@@ -1,18 +1,13 @@
 namespace AddressRegistry.Projections.Legacy
 {
-    using System;
-    using System.IO;
     using AddressDetail;
     using AddressList;
     using AddressMatch;
     using AddressSyndication;
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.Runner;
-    using Be.Vlaanderen.Basisregisters.ProjectionHandling.Runner.MigrationExtensions;
     using CrabIdToPersistentLocalId;
     using Infrastructure;
     using Microsoft.EntityFrameworkCore;
-    using Microsoft.EntityFrameworkCore.Design;
-    using Microsoft.Extensions.Configuration;
 
     public class LegacyContext : RunnerDbContext<LegacyContext>
     {
@@ -33,34 +28,5 @@ namespace AddressRegistry.Projections.Legacy
         // This needs to be DbContextOptions<T> for Autofac!
         public LegacyContext(DbContextOptions<LegacyContext> options)
             : base(options) { }
-    }
-
-    public class ConfigBasedContextFactory : IDesignTimeDbContextFactory<LegacyContext>
-    {
-        public LegacyContext CreateDbContext(string[] args)
-        {
-            const string migrationConnectionStringName = "LegacyProjectionsAdmin";
-
-            var configuration = new ConfigurationBuilder()
-                .SetBasePath(Directory.GetCurrentDirectory())
-                .AddJsonFile("appsettings.json", optional: false, reloadOnChange: false)
-                .AddJsonFile($"appsettings.{Environment.MachineName.ToLowerInvariant()}.json", optional: true, reloadOnChange: false)
-                .AddEnvironmentVariables()
-                .Build();
-
-            var connectionString = configuration.GetConnectionString(migrationConnectionStringName);
-            if (string.IsNullOrEmpty(connectionString))
-                throw new InvalidOperationException($"Could not find a connection string with name '{migrationConnectionStringName}'");
-
-            var builder = new DbContextOptionsBuilder<LegacyContext>()
-                .UseSqlServer(connectionString, sqlServerOptions =>
-                {
-                    sqlServerOptions.EnableRetryOnFailure();
-                    sqlServerOptions.MigrationsHistoryTable(MigrationTables.Legacy, Schema.Legacy);
-                })
-                .UseExtendedSqlServerMigrations();
-
-            return new LegacyContext(builder.Options);
-        }
     }
 }