+++ src/AddressRegistry.Projections.Legacy/LegacyContextMigrationFactory.cs
@@ -3,13 +3,11 @@ namespace AddressRegistry.Projections.Legacy
     using Be.Vlaanderen.Basisregisters.ProjectionHandling.Runner;
     using Infrastructure;
     using Microsoft.EntityFrameworkCore;
-    using Microsoft.EntityFrameworkCore.Infrastructure;
 
     public class LegacyContextMigrationFactory : RunnerDbContextMigrationFactory<LegacyContext>
     {
         public LegacyContextMigrationFactory()
-            : base("LegacyProjectionsAdmin", HistoryConfiguration)
-        { }
+            : base("LegacyProjectionsAdmin", HistoryConfiguration) { }
 
         private static MigrationHistoryConfiguration HistoryConfiguration =>
             new MigrationHistoryConfiguration
@@ -20,8 +18,5 @@ namespace AddressRegistry.Projections.Legacy
 
         protected override LegacyContext CreateContext(DbContextOptions<LegacyContext> migrationContextOptions)
             => new LegacyContext(migrationContextOptions);
-
-        //protected override void ConfigureSqlServerOptions(SqlServerDbContextOptionsBuilder sqlServerOptions)
-        //    => sqlServerOptions.UseNetTopologySuite();
     }
 }